automated-momentum-trading-experiment/requirements.txt
python-deriv-api==0.1.3
python-dotenv==1.0.0
reactivex==4.0.4
typing_extensions==4.5.0
websockets==10.3
